Move cast_value to component method
