+++ omc4py/classes.py
@@ -313,6 +313,12 @@ class Component(
         return Component(self.class_, tuple(dimensions_generator()))
 
     @property
+    def is_array(self) -> bool: return bool(self.dimensions)
+
+    @property
+    def is_scalar(self) -> bool: return not self.is_array
+
+    @property
     def class_restrictions(
         self,
     ) -> typing.Tuple[typing.Type, ...]:
@@ -330,6 +336,94 @@ class Component(
             return tuple({VariableName, str})
         else:
             return ()
+
+    def cast_value(
+        self,
+        name: str,
+        value: typing.Any,
+        required: REQUIRED_or_OPTIONAL = "required",
+    ) -> typing.Any:
+        if value is None:
+            if required == "required":
+                raise ValueError(
+                    f"{name!r} must not be None"
+                )
+            if required == "optional":
+                return None
+            else:
+                raise ValueError(
+                    f"required must be (required|optional) got {required!r}"
+                )
+
+        if self.is_scalar:
+            class_restrictions = self.class_restrictions
+            if class_restrictions:
+                if not isinstance(value, class_restrictions):
+                    raise TypeError(
+                        f"{name!r} must be an instance of {class_restrictions}"
+                        f", got {value!r}: {type(value)!r}"
+                    )
+            return self.class_(value)
+
+        else:  # array
+            return self.__cast_array_value(
+                name,
+                value,
+            )
+
+    def __cast_array_value(
+        self,
+        name: str,
+        value: typing.Any,
+    ) -> numpy.ndarray:
+        object_array = numpy.array(value, dtype=object)
+
+        same_n_dimensions = (len(self.dimensions) == object_array.ndim)
+        dimensions_are_correct = [
+            True if expected is None else expected == actual
+            for expected, actual in zip(self.dimensions, object_array.shape)
+        ]
+
+        if not(same_n_dimensions and all(dimensions_are_correct)):
+            raise ValueError(
+                "Shape of the array "
+                f"must be {self.dimensions_to_str(self.dimensions)}, "
+                f"got {self.dimensions_to_str(object_array.shape)}"
+            )
+
+        class_restrictions = self.class_restrictions
+        if class_restrictions:
+            isinstance_vectorized = numpy.vectorize(
+                lambda cls: isinstance(cls, class_restrictions),
+                otypes=[numpy.dtype(bool)],
+            )
+            isinstance_mask = isinstance_vectorized(
+                object_array
+            )
+            if not numpy.all(isinstance_mask):
+                raise TypeError(
+                    f"All items of the array {name!r} "
+                    f"must be instances of {class_restrictions}"
+                )
+
+        class_vectorized = numpy.vectorize(
+            self.class_,
+            otypes=[numpy.dtype(self.class_)]
+        )
+        return class_vectorized(object_array)
+
+    @staticmethod
+    def dimensions_to_str(
+        sizes: Dimensions,
+    ) -> str:
+        return (
+            "{"
+            + ", ".join(
+                str(size) if size is not None else ":"
+                for size in sizes
+            )
+            + "}"
+        )
 
 
 class AbstractOMCInteractive(
@@ -583,9 +677,8 @@ class ModelicaRecordMeta(
                 )
 
             for key, value in self.__dict.items():
-                self.__dict[key] = compiler.cast_value(
-                    elements[key], key, value
-                )
+                component = typing.cast(Component, elements[key])
+                self.__dict[key] = component.cast_value(key, value)
 
             return self
 
@@ -846,5 +939,4 @@ class ModelicaFunction(
     ...
 
 
-from . import compiler  # noqa: E402
 from . import parser  # noqa: E402
